Moved core/command to internal/app/command
