+++ cmd/core-command/toml_test.go
@@ -10,7 +10,7 @@ package main
 import (
 	"testing"
 
-	"github.com/edgexfoundry/edgex-go/core/command"
+	"github.com/edgexfoundry/edgex-go/internal/app/command"
 	"github.com/edgexfoundry/edgex-go/internal/pkg/config"
 )
