+++ internal/app/command/device.go
