+++ internal/app/command/raml/core-command.raml
