+++ internal/app/command/rest.go
