+++ internal/app/command/restDevice.go
